Rename helper.py to utils.py
Rename and refactor helper.py to utils.py to make naming more
consistent with established practices.
