+++ ciscosparkapi/api/messages.py
@@ -15,7 +15,7 @@ from six import string_types
 from requests_toolbelt import MultipartEncoder
 
 from ciscosparkapi.exceptions import ciscosparkapiException
-from ciscosparkapi.helper import generator_container, is_web_url, \
+from ciscosparkapi.utils import generator_container, is_web_url, \
     is_local_file, open_local_file
 from ciscosparkapi.restsession import RestSession
 from ciscosparkapi.sparkdata import SparkData
