+++ ciscosparkapi/utils.py
